Fourth Commit
dbUser.js <-> Added New Collection and Mongoose Schema from storing userData.
server.js <-> Made a post receiver to find ONE doc based on the userId and return.
TweetBox.js <-> Add functionality to  check if userData is on the localStorage. If there is none, then make a request to fetch and store